+++ src/main/java/subway/service/line/LineService.java
@@ -1,4 +1,4 @@
-package subway.service;
+package subway.service.line;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
