+++ src/main/java/subway/service/section/SectionService.java
@@ -1,4 +1,4 @@
-package subway.service;
+package subway.service.section;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
