+++ src/main/java/subway/ui/LineController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
-import subway.service.LineService;
+import subway.service.line.LineService;
 
 import javax.validation.Valid;
 import java.net.URI;
