+++ src/test/java/subway/service/SectionServiceTest.java
@@ -12,6 +12,7 @@ import subway.domain.Sections;
 import subway.dto.LineStationRequest;
 import subway.dto.LineStationResponse;
 import subway.persistence.repository.SectionRepository;
+import subway.service.section.SectionService;
 
 import java.util.ArrayList;
 import java.util.List;
